backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0
python-dotenv==1.0.0
pillow==10.1.0
qrcode[pil]==7.4.2
geopy==2.4.1
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy>=1.21.0
